python/device/requirements.txt
inflection
opencensus-ext-azure
azure-iot-device
azure-iot-hub

